+++ api-service/src/main/java/com/code/ecommercebackend/services/common/CommonFunction.java
@@ -4,5 +4,5 @@ public interface CommonFunction {
     void saveUserBehavior(String token,
                           long behavior,
                           long productId,
-                          Integer rating);
+                          Float rating);
 }
